Revert "Revert "PPF-1227: dedicated ap module""
