+++ PyPDFForm/wrapper.py
@@ -24,6 +24,7 @@ from functools import cached_property
 from typing import TYPE_CHECKING, BinaryIO, Dict, List, Sequence, Tuple, Union
 
 from .adapter import fp_or_f_obj_or_stream_to_stream
+from .ap import appearance_streams_handler
 from .constants import (DEFAULT_FONT, DEFAULT_FONT_COLOR, DEFAULT_FONT_SIZE,
                         VERSION_IDENTIFIER_PREFIX, VERSION_IDENTIFIERS)
 from .coordinate import generate_coordinate_grid
@@ -37,8 +38,8 @@ from .middleware.dropdown import Dropdown
 from .middleware.signature import Signature
 from .middleware.text import Text
 from .template import build_widgets, update_widget_keys
-from .utils import (appearance_streams_handler, generate_unique_suffix,
-                    get_page_streams, merge_two_pdfs, remove_all_widgets)
+from .utils import (generate_unique_suffix, get_page_streams, merge_two_pdfs,
+                    remove_all_widgets)
 from .watermark import (copy_watermark_widgets, create_watermarks_and_draw,
                         merge_watermarks_with_pdf)
 from .widgets import CheckBoxField, ImageField, RadioGroup, SignatureField
